requirements.txt
httpx
fastapi
uvicorn[standard]
pydantic
sqlalchemy
aiosqlite
sqlmodel
playwright
nats-py